refactor(chat): move helpers to utils/ package and remove stream_chat_response wrapper
- Rename helpers.py → routes/utils/__init__.py so utilities import as `from ...utils`
- Extract ChatStreamProcessor into routes/utils/stream_processor.py alongside
  streaming_error_message (now public)
- Remove the stream_chat_response thin wrapper; chat_completion now instantiates
  ChatStreamProcessor directly
- Update all tests and patch paths to reflect the new module layout
