+++ app/core_plugins/chat/tests/test_rag_integration.py
@@ -7,7 +7,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 from fastapi import HTTPException
 
-from app.core_plugins.chat.routes.helpers import extract_query_text, resolve_document_blocks
+from app.core_plugins.chat.routes.utils import extract_query_text, resolve_document_blocks
 from app.core_plugins.chat.schemas import ChatMessage
 from app.lib.rag import (
     DocumentNotFoundError,
@@ -16,7 +16,7 @@ from app.lib.rag import (
     RetrievedChunk,
 )
 
-RETRIEVE_CONTEXT_PATH = "app.core_plugins.chat.routes.helpers.agentic_retrieve_context"
+RETRIEVE_CONTEXT_PATH = "app.core_plugins.chat.routes.utils.agentic_retrieve_context"
 
 
 def _user_msg(content: str | list[Any]) -> ChatMessage:
