+++ app/core_plugins/chat/tests/test_scope_validation.py
@@ -14,7 +14,7 @@ from app.core.config import get_settings
 from app.core.encryption import get_encryption_service
 from app.core_plugins.chat.models import Conversation
 from app.core_plugins.chat.prompt import REFUSAL_MESSAGE, is_query_in_scope
-from app.core_plugins.chat.routes.helpers import stream_out_of_scope_refusal
+from app.core_plugins.chat.routes.utils import stream_out_of_scope_refusal
 from app.core_plugins.chat.schemas import ChatCompletionResponse, ChatMessage
 from app.models.llm import LLMConfig
 from app.models.user import User
@@ -164,7 +164,7 @@ class TestOutOfScopeConversationCreation:
 
         with (
             patch("app.core_plugins.chat.routes.completions.get_provider"),
-            patch("app.core_plugins.chat.routes.helpers.is_query_in_scope", return_value=False),
+            patch("app.core_plugins.chat.routes.utils.is_query_in_scope", return_value=False),
             patch(
                 "app.core_plugins.chat.service.ChatService.add_message",
                 new_callable=AsyncMock,
@@ -212,7 +212,7 @@ class TestOutOfScopeConversationCreation:
 
         with (
             patch("app.core_plugins.chat.routes.completions.get_provider"),
-            patch("app.core_plugins.chat.routes.helpers.is_query_in_scope", return_value=False),
+            patch("app.core_plugins.chat.routes.utils.is_query_in_scope", return_value=False),
             patch(
                 "app.core_plugins.chat.service.ChatService.get_conversation_messages",
                 new_callable=AsyncMock,
